ORDEN DE MATERIA PRIMA: Funcionalidad completa finalizada e integrada OK. (persistencia,generacion de pdf con distinto nombre, busq de proveedor, busq de materia prima)
MONITOR: FUNCIONALIDAD FINALIZADA/ FALTA PARTE DE LA INTEGRACION DEBIDO AL PROBLEMA QUE LES COMENTE a la hora de querer agregar un item a una orden de pedido de comida,/ 

Observaciones> Solo me quedaria mejorar el disenio visual de las tablas, para que quede mas lindo.
IMPORTANTE> Me arroja el siguiente error>

The constructor Comanda() is undefined	seleccionDeCliente.java	/Sistema Pizzeria Sin Persistencia/src/presentacion/vista

The method generarTicket() in the type Ticket is not applicable for the arguments (PedidoDTO)	seleccionDeCliente.java	/Sistema Pizzeria Sin Persistencia/src/presentacion/vista

The method generarComanda() in the type Comanda is not applicable for the arguments (PedidoDTO)	seleccionDeCliente.java	/Sistema Pizzeria Sin Persistencia/src/presentacion/vista

The constructor Ticket() is undefined	seleccionDeCliente.java	/Sistema Pizzeria Sin Persistencia/src/presentacion/vista

+++ src/dao/PedidoDAO.java
@@ -5,16 +5,11 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.GregorianCalendar;
-import java.util.Iterator;
 import java.util.List;
 
 import modelo.Clientes;
 import modelo.Items;
 import modelo.Ofertas;
-import modelo.Pedidos;
-
 import dto.ClienteDTO;
 import dto.ItemDTO;
 import dto.OfertaDTO;
@@ -174,7 +169,7 @@ public class PedidoDAO
 				resultSet.getString("fecha"),resultSet.getString("hora"),estadoPedido,
 				resultSet.getInt("total"),resultSet.getInt("ticket"),
 				resultSet.getInt("comanda"),
-				ClienteDTO.buscarCliente(cli.obtenerClientes(),(Integer)resultSet.getInt("cliente")),
+				ClienteDTO.buscarCliente(cli.obtenerClientes(),resultSet.getInt("cliente")),
 				resultSet.getBoolean("llevadelivery"),listOfertas,resultSet.getBoolean("fueeliminado"));
 				pedidos.add(aux);
 			}
